@@ -25,17 +25,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
+    "@angular/animations": "^17.3.12",
     "@angular/cdk": "^16.2.12",
-    "@angular/common": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^16.2.12",
-    "@angular/localize": "^16.2.12",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/platform-server": "^16.2.12",
-    "@angular/router": "^16.2.12",
+    "@angular/common": "^17.3.12",
+    "@angular/compiler": "^17.3.12",
+    "@angular/core": "^17.3.12",
+    "@angular/forms": "^17.3.12",
+    "@angular/localize": "^17.3.12",
+    "@angular/platform-browser": "^17.3.12",
+    "@angular/platform-browser-dynamic": "^17.3.12",
+    "@angular/platform-server": "^17.3.12",
+    "@angular/router": "^17.3.12",
     "@microsoft/applicationinsights-web": "^2.5.10",
     "@ng-bootstrap/ng-bootstrap": "^15.0.0",
     "@ng-select/ng-select": "^11.0.0",
@@ -70,30 +70,35 @@
     "tinymce": "^6.4.2",
     "tslib": "^2.3.1",
     "webpack": "^5.30.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.13",
-    "@angular-eslint/eslint-plugin": "^12.6.1",
-    "@angular/cli": "^16.2.10",
-    "@angular/compiler-cli": "^16.2.12",
-    "@angular/language-service": "^16.2.12",
+    "@angular-devkit/build-angular": "^17.3.11",
+    "@angular-eslint/builder": "^17.0.0",
+    "@angular-eslint/eslint-plugin": "^17.0.0",
+    "@angular-eslint/eslint-plugin-template": "^17.0.0",
+    "@angular-eslint/schematics": "^17.0.0",
+    "@angular-eslint/template-parser": "^17.0.0",
+    "@angular/cli": "^17.3.11",
+    "@angular/compiler-cli": "^17.3.12",
+    "@angular/language-service": "^17.3.12",
     "@babel/compat-data": "~7.9.0",
     "@fortawesome/fontawesome-free": "^5.13.0",
     "@openapitools/openapi-generator-cli": "^2.4.26",
     "@types/node": "^13.13.50",
-    "@typescript-eslint/eslint-plugin": "^5.3.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.3.0",
-    "@typescript-eslint/parser": "^5.3.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@typescript-eslint/types": "^6.19.0",
+    "eslint": "^8.51.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jsdoc": "^37.0.3",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "install": "^0.13.0",
-    "npm": "^10.2.5",
+    "npm": "^10.6.0",
     "protractor": "^7.0.0",
     "ts-node": "~8.9.1",
-    "typescript": "4.9.5"
+    "typescript": "^5.2.0"
   }
 }
