@@ -2,9 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ICellRendererAngularComp } from 'ag-grid-angular';
 
 @Component({
-  selector: 'hippocamp-custom-pinned-row-renderer',
-  templateUrl: './custom-pinned-row-renderer.component.html',
-  styleUrls: ['./custom-pinned-row-renderer.component.scss']
+    selector: 'hippocamp-custom-pinned-row-renderer',
+    templateUrl: './custom-pinned-row-renderer.component.html',
+    styleUrls: ['./custom-pinned-row-renderer.component.scss'],
+    standalone: true
 })
 export class CustomPinnedRowRendererComponent implements ICellRendererAngularComp {
 
