@@ -1,10 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { AgRendererComponent } from 'ag-grid-angular';
+import { RouterLink } from '@angular/router';
+import { NgIf } from '@angular/common';
 
 @Component({
-  selector: 'hippocamp-fontawesome-icon-link-renderer',
-  templateUrl: './fontawesome-icon-link-renderer.component.html',
-  styleUrls: ['./fontawesome-icon-link-renderer.component.scss']
+    selector: 'hippocamp-fontawesome-icon-link-renderer',
+    templateUrl: './fontawesome-icon-link-renderer.component.html',
+    styleUrls: ['./fontawesome-icon-link-renderer.component.scss'],
+    standalone: true,
+    imports: [NgIf, RouterLink]
 })
 export class FontAwesomeIconLinkRendererComponent implements AgRendererComponent {
   params: any;    
