@@ -1,10 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { AgRendererComponent } from 'ag-grid-angular';
+import { RouterLink } from '@angular/router';
+import { NgIf } from '@angular/common';
 
 @Component({
-  selector: 'hippocamp-link-renderer',
-  templateUrl: './link-renderer.component.html',
-  styleUrls: ['./link-renderer.component.scss']
+    selector: 'hippocamp-link-renderer',
+    templateUrl: './link-renderer.component.html',
+    styleUrls: ['./link-renderer.component.scss'],
+    standalone: true,
+    imports: [NgIf, RouterLink]
 })
 
 export class LinkRendererComponent implements AgRendererComponent {
