@@ -1,10 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { AgRendererComponent } from 'ag-grid-angular';
+import { RouterLink } from '@angular/router';
+import { NgFor, NgIf } from '@angular/common';
 
 @Component({
-  selector: 'hippocamp-multi-link-renderer',
-  templateUrl: './multi-link-renderer.component.html',
-  styleUrls: ['./multi-link-renderer.component.scss']
+    selector: 'hippocamp-multi-link-renderer',
+    templateUrl: './multi-link-renderer.component.html',
+    styleUrls: ['./multi-link-renderer.component.scss'],
+    standalone: true,
+    imports: [NgFor, NgIf, RouterLink]
 })
 
 export class MultiLinkRendererComponent implements AgRendererComponent {
