@@ -2,11 +2,14 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Alert } from '../../models/alert';
 import { AlertContext } from '../../models/enums/alert-context.enum';
 import { AlertService } from '../../services/alert.service';
+import { NgIf } from '@angular/common';
 
 @Component({
-  selector: 'alert',
-  templateUrl: './alert.component.html',
-  styleUrls: ['./alert.component.scss']
+    selector: 'alert',
+    templateUrl: './alert.component.html',
+    styleUrls: ['./alert.component.scss'],
+    standalone: true,
+    imports: [NgIf]
 })
 export class AlertComponent implements OnInit {
 
